Refactoring shape checking to reduce redundant checks (MakeShape now takes an int64 instead of an int, avoiding some of the casting ugliness and reducing the need for callers to do their own, redundant checks). Fixing additional int32->64 warnings Change: 121498517

+++ tensorflow/core/framework/tensor_shape.h
@@ -265,8 +265,8 @@ class TensorShapeUtils {
 
   /// \brief Returns a `TensorShape` whose dimensions are
   /// `dims[0]`, `dims[1]`, ..., `dims[n-1]`.
-  static Status MakeShape(const int32* dims, int n, TensorShape* out);
-  static Status MakeShape(const int64* dims, int n, TensorShape* out);
+  static Status MakeShape(const int32* dims, int64 n, TensorShape* out);
+  static Status MakeShape(const int64* dims, int64 n, TensorShape* out);
 
   static string ShapeListString(const gtl::ArraySlice<TensorShape>& shapes);
